requirements.txt
Flask
Flask-SQLAlchemy
Flask-Injector
psycopg2-binary
python-dotenv
pydantic-settings
pydantic